added CrimeMap.js component to make code cleaner.
